+++ src/command/default/SayCommand.js
@@ -3,18 +3,18 @@ const Command = require("../Command");
 const ConsoleCommandSender = require("../ConsoleCommandSender");
 
 class SayCommand extends Command {
-    constructor(){
+    constructor() {
         super("say", "broadcast a message to the players");
     }
 
-    execute(sender, args){
+    execute(sender, args) {
         if(args.length < 1){
             if(sender instanceof ConsoleCommandSender){
                 sender.sendMessage(TextFormat.RED + "/say <message>");
             }
             return;
         }
-        sender.getServer().broadcastMessage(TextFormat.RED + "[CONSOLE] " + TextFormat.WHITE + args.join(" "));
+        sender.getServer().broadcastMessage(TextFormat.DARK_PURPLE + "[Server] " + TextFormat.WHITE + args.join(""));
     }
 }